+++ sys/src/9/port/proc.c
@@ -912,7 +912,7 @@ found:
 
 	p->state = Scheding;
 	hi("runproc, pm->mp = sys->machptr[m->machno]\n");
-	hi("runproc, sys->machptr[m->machno] = "); put64(&sys->machptr[m->machno]); hi("\n");
+	hi("runproc, sys->machptr[m->machno] = "); put64((uint64_t)&sys->machptr[m->machno]); hi("\n");
 	p->mp = sys->machptr[m->machno];
 
 	if(edflock(p)){
